[ADD] 문제 채점 결과를 얻어올 수 있는 API 제작함
1. 문제 채점 결과를 얻어올 수 있는 API 제작함
2. DB 및 SQL 문 수정함

+++ problem-service/src/main/java/com/adele/problemservice/dto/KafkaCompile.java
@@ -14,7 +14,7 @@ import java.io.Serializable;
 @ToString
 public class KafkaCompile implements Serializable {
     private CompileStatus compileStatus;
-    private int caseNo;
+    private Long caseNo;
     private String inputSource;
     private String outputSource;
     private CompileErrorReason compileErrorReason;
